refactor: delimiter handelr 클래스 삭제 - 제거한 클래스 역할이 customDelimiterExtractor와 같아 로직 customDelimiterExtractor에 합침 - 의존성 변경에 따른 appConfig 및 stringSplitter 클래스 변경
